Consolidate node and client db/submitToChain fns
